correct cycle #358, by extracting TransformationType in a separate file
because:
    * TransformationType enum is used by both TransformOperationExecutor  and MetadataStorage
    * TransformationType was in the same file as TransformOperationExecutor
    * TransformOperationExecutor require MetadataStorage
so: importing TransformationType from TransformOperationExecutor.ts was as a side effect importing MetadataStorage.ts that was importing TransformationType etc. creating a cycle
